shared_mem: move schemas + numpy golden into hist.py (data-model home)
Relocate the data model out of hist_demo.py into hist.py (matching simp_fun.py's
convention of the component file owning its schemas): HistCmd / HistResp /
HistError / HistErrorField, the type fields (Tx/Ndata/Nbin/Uint32/Float32/Addr),
the MAX_*/MEM_*/STREAM_DWIDTH/INCLUDE_DIR constants, SCHEMA_CLASSES, and the numpy
golden HistogramAccel + HistSimResult.

This is prep for retiring hist_demo.py: hist.py had imported these FROM hist_demo,
which would have made hist_build.py (importing HistAccel from hist.py) circular.
Now hist.py is the base; hist_demo.py re-exports the moved symbols (same objects)
so every `from hist_demo import <schema>` consumer keeps working until hist_demo
is retired in Phase 5. No behavior change — non-vitis suite green aside from the
known pre-existing failures.
